task11完了

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -12,39 +12,47 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import jp.co.seattle.library.dto.BookRentInfo;
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentBookService;
 
 /**
  * 削除コントローラー
  */
-@Controller //APIの入り口
+@Controller // APIの入り口
 public class DeleteBookController {
-    final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
-    
-    @Autowired
+	final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
+
+	@Autowired
 	private BooksService booksService;
-    
-    /**
-     * 対象書籍を削除する
-     *
-     * @param locale ロケール情報
-     * @param bookId 書籍ID
-     * @param model モデル情報
-     * @return 遷移先画面名
-     */
-    @Transactional
-    @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
-    public String deleteBook(
-            Locale locale,
-            @RequestParam("bookId") Integer bookId,
-            Model model) {
-        logger.info("Welcome delete! The client locale is {}.", locale);
+	@Autowired
+	private RentBookService rentBookService;
 
-        booksService.deleteBook(bookId);
-        
-        model.addAttribute("bookList", booksService.getBookList());
-        return "home";
+	/**
+	 * 対象書籍を削除する
+	 *
+	 * @param locale ロケール情報
+	 * @param bookId 書籍ID
+	 * @param model  モデル情報
+	 * @return 遷移先画面名
+	 */
+	@Transactional
+	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
+	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model) {
+		logger.info("Welcome delete! The client locale is {}.", locale);
 
-    }
-   
+		BookRentInfo selecteddeletebookInfo = rentBookService.selectRentBookInfo(bookId);
+
+		if (selecteddeletebookInfo != null) {
+			model.addAttribute("errorMessage", "貸出し中のため削除できません。");
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+		} else {
+
+			booksService.deleteBook(bookId);
+
+			model.addAttribute("bookList", booksService.getBookList());
+			return "home";
+		}
+	}
 }
